Move navAggregate algorithm I/O structs into the algorithm header
